requirements.txt
disnake
mafic
aiohttp
python-dotenv
requests